refactor(core): centralize blocked path segment verification
